+++ src/main/java/pack/mapper/kakao/KakaoMapper.java
@@ -1,4 +1,4 @@
-package pack.repository.kakao;
+package pack.mapper.kakao;
 
 import java.util.List;
