+++ src/main/resources/mybatis/mapper/mapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="pack.repository.data.DataMapping">
+<mapper namespace="pack.mapper.data.DataMapping">
 	<select id="selectList" resultType="pack.dto.board.BoardDTO">
 		select * from notice_board order by bdate desc
 	</select>
